uint64

+++ internal/verifier/api/change_event_counts.go
@@ -5,22 +5,22 @@ import "github.com/rs/zerolog"
 // ChangeEventCounts tallies cumulative change events seen by a change reader,
 // across all generations, since the verifier first started.
 type ChangeEventCounts struct {
-	Insert  int
-	Update  int
-	Replace int
-	Delete  int
+	Insert  uint64
+	Update  uint64
+	Replace uint64
+	Delete  uint64
 }
 
 var _ zerolog.LogObjectMarshaler = ChangeEventCounts{}
 
-func (cec ChangeEventCounts) Total() int {
+func (cec ChangeEventCounts) Total() uint64 {
 	return cec.Insert + cec.Update + cec.Replace + cec.Delete
 }
 
 func (cec ChangeEventCounts) MarshalZerologObject(e *zerolog.Event) {
 	e.
-		Int("insert", cec.Insert).
-		Int("update", cec.Update).
-		Int("replace", cec.Replace).
-		Int("delete", cec.Delete)
+		Uint64("insert", cec.Insert).
+		Uint64("update", cec.Update).
+		Uint64("replace", cec.Replace).
+		Uint64("delete", cec.Delete)
 }
